requirements.txt
requests>=2.31
beautifulsoup4>=4.12
pandas>=2.2
lxml>=5.1
neo4j>=5.17
psycopg2-binary>=2.9
openai>=1.12
python-dotenv>=1.0
tqdm>=4.66
